feat: allow batching of simulatetx
Extends the `.batch` method of the wallet interface to support `.simulateTx`, since there's no good reason not to and it enables more meaningful interactions.

Also tweaked `BatchCall` to use it, so the feature is immediately available for the ecosystem to use it...but again, it's probably going to be dead soon.

Co-authored-by: thunkar <gregojquiros@gmail.com>